Class modification (Vid)
Lots of changes in the Vid class: functions to create and save videos
are now methods of Vid. 'step', 'scale' & 'sizeOut' are properties of
Vid for code simplification. 2 news properties (frameSc & frameOut) to
store Img array to make computing faster and to make easier test and
debugging. Constructor of vid more complete (includes the path and
video creation).